backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
